refactor(intents): rename capture_event package to captureevent
Fix ST1003 (underscore in package name), ST1000, and revive
package-comments lint violations:

- Rename directory intents/capture_event/ to intents/captureevent/
- Update package declaration from capture_event to captureevent in
  all 14 source and test files
- Rename suite test file to captureevent_suite_test.go
- Update package doc comment to reference captureevent
- Remove now-redundant import aliases in 4 external files (app.go,
  registrar.go, helpers.go, duplication_test.go) since the package
  name matches the existing alias

AI-Generated-By: OpenCode (Claude Sonnet 4)
