+++ internal/cli/intents/captureevent/benchmarks_test.go
@@ -1,4 +1,4 @@
-package capture_event
+package captureevent
 
 import (
 	"testing"
