+++ internal/cli/intents/captureevent/global_keys_test.go
@@ -1,4 +1,4 @@
-package capture_event
+package captureevent
 
 import (
 	"time"
