+++ internal/cli/intents/captureevent/handlers.go
@@ -1,4 +1,4 @@
-package capture_event
+package captureevent
 
 import (
 	"context"
