+++ internal/cli/intents/captureevent/helpers.go
@@ -1,4 +1,4 @@
-package capture_event
+package captureevent
 
 import (
 	"context"
