+++ internal/cli/intents/captureevent/intent.go
@@ -1,4 +1,4 @@
-package capture_event
+package captureevent
 
 import (
 	"time"
