+++ internal/cli/intents/captureevent/workflow_e2e_test.go
@@ -1,4 +1,4 @@
-package capture_event_test
+package captureevent_test
 
 import (
 	"strings"
